[fix]高さと重さの単位設定に誤りがあったため修正

+++ app/src/main/java/com/example/pokebook/ui/viewModel/PokemonDetailState.kt
@@ -36,8 +36,8 @@ data class PokemonDetailScreenUiData(
     val defense: Int = 0,
     val speed: Int = 0,
     val imageUri: String = "",
-    val height: Int = 0,
-    val weight: Int = 0
+    val height: Double = 0.0,
+    val weight: Double = 0.0
 )
 
 /**
